move add_tel_link to _dependencies/commons
