+++ src/communicate/_utils/handlers/button_handlers.py
@@ -10,9 +10,8 @@ from telegram import (
     WebAppInfo,
 )
 
-from _dependencies.commons import get_app_config
+from _dependencies.commons import get_app_config, add_tel_link
 from _dependencies.pubsub import notify_admin
-from compose_notifications._utils.commons import add_tel_link
 
 from ..buttons import (
     Commands,
